pyproject.toml
[project]
name = "Hybrid_RAG"
version = "0.1.0"
description = "DOST Chatbot"
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    "jupyter>=1.1.1",
    "langchain-community>=0.3.26",
    "langchain-core>=0.3.67",
    "langchain-openai>=0.3.27",
    "langchain-qdrant>=0.2.0",
    "langgraph>=0.5.0",
    "langsmith>=0.4.4",
    "nltk>=3.9.1",
    "pandas>=2.3.0",
    "pillow>=11.3.0",
    "pymupdf>=1.26.1",
    "ragas>=0.3.0",
    "rapidfuzz>=3.13.0",
    "tavily_python>=0.3.4",
    "python-dotenv>=1.0.1",
    "streamlit>=1.38.0",
]
